Changed data type of seed from unsigned long long to unsigned long to comply with OpenCL

+++ src/cl_mc_kernel.hpp
@@ -9,7 +9,7 @@ class cl_mc_kernel : public mc_kernel
 {
 public:
 	// Initialize kernel.
-	explicit cl_mc_kernel(const float* h_sf_e, const float* h_sf_d, const int h_sf_ns, const int h_sf_ne, const float* h_corner0, const float* h_corner1, const int* h_num_probes, const float h_granularity_inverse, const int num_mc_tasks, const int h_ng, const unsigned long long h_seed);
+	explicit cl_mc_kernel(const float* h_sf_e, const float* h_sf_d, const int h_sf_ns, const int h_sf_ne, const float* h_corner0, const float* h_corner1, const int* h_num_probes, const float h_granularity_inverse, const int num_mc_tasks, const int h_ng, const unsigned long h_seed);
 
 	// Update grid map pointers.
 	void update(const vector<vector<float> > h_maps, const vector<size_t>& xs);
